support mardown titles and keep the untouched title in the decomposed file
Mardown "#" titles are supported as "----" separators, and even
preferred over them.

Also, the title of a section is put inside the text file and read from
there. If the title was only in the filename, it would have been
mangled:
"foo - bar" -> "foo-bar.txt" -> "foo bar".
Bonus: the decomposed text is pretty markdown file with a title.