change argument path logic to avoid mistakes and accept --merge/--split
If arguments are in/out rather than file/dir, the arguments order will
be different when -c or -d is given, and the correct argument order can
still be verified easily.
This way, the -c and -d cannot be given interchangeabily, which can
avoid some accidental bad operation choice, as the in/out argument order
is more natural and less easy to be mistaken.

Also, new forms are accepted, --merge and --split.